requirements.txt
pandas
streamlit
scikit-learn
google-generativeai
joblib